change styling, refactor
