+++ client/components/chat/ChatScreen.js
@@ -1,133 +1,13 @@
-import { useState, useEffect, useRef } from "react";
-import { BsThreeDotsVertical, BsSearch } from "react-icons/bs";
-import Image from "next/image";
-import Dropdown from "../Dropdown";
-import Message from "./Message";
-import { useSelector } from "react-redux";
-import { selectActiveChat } from "../../features/chatSlice";
+import ChatHeader from "./ChatHeader";
+import Messages from "./Messages";
+import MessageInput from "./MessageInput";
 
 const ChatScreen = ({ messageHistory, sendMessage }) => {
-  const activeChat = useSelector(selectActiveChat);
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-  const [message, setMessage] = useState("");
-  const messageRef = useRef(null);
-
-  const openDropdownHandler = () => {
-    setIsDropdownOpen((prevState) => !prevState);
-  };
-
-  const handleChangeMessage = (e) => {
-    setMessage(e.target.value);
-  };
-
-  const onKeyUpHandler = (e) => {
-    if (e.key === "Enter" || e.key === "NumpadEnter") {
-      sendMessageHandler(e);
-    }
-  };
-
-  const sendMessageHandler = async (e) => {
-    e.preventDefault();
-    sendMessage({
-      type: "chat_message",
-      message,
-    });
-    setMessage("");
-  };
-
-  useEffect(() => {
-    if (messageRef && messageRef.current) {
-      messageRef.current.style.height = "0px";
-      const scrollHeight = messageRef.current.scrollHeight;
-      messageRef.current.style.height = scrollHeight + "px";
-    }
-  }, [message]);
-
-  const actions = [
-    {
-      action: "Add to conversation",
-      func: () => console.log("Add to conversation"),
-    },
-    {
-      action: "Settings",
-      func: () => console.log("Settings"),
-    },
-    {
-      action: "Leave",
-      func: () => console.log("Leave"),
-    },
-  ];
-
   return (
-    <div className="flex flex-col w-full lg:w-[calc(100%-400px)] h-full bg-primary self-end">
-      {/* header */}
-      <div className="flex p-5 w-full h-20 justify-between items-center bg-secondary text-stone-100">
-        <div className="flex space-x-2.5 items-center">
-          <div className="w-10 h-10 rounded-full cursor-pointer">
-            <Image
-              layout="responsive"
-              width="100%"
-              height="100%"
-              className="rounded-full object-cover"
-              src="/avatar.png"
-              alt="avatar"
-            />
-          </div>
-          <div>
-            <p className="text-base font-semibold">{activeChat.name || ""}</p>
-          </div>
-        </div>
-        <div className="flex space-x-2.5">
-          <div className="p-2 rounded-full cursor-pointer">
-            <span className="text-2xl">
-              <BsSearch />
-            </span>
-          </div>
-          <div
-            className={`relative p-2 rounded-full cursor-pointer ${
-              isDropdownOpen ? "bg-primary" : null
-            }`}
-            onClick={openDropdownHandler}
-          >
-            <span className="text-2xl">
-              <BsThreeDotsVertical />
-            </span>
-            <Dropdown actions={actions} isOpen={isDropdownOpen} />
-          </div>
-        </div>
-      </div>
-
-      {/* messages */}
-      <div className="flex flex-col grow w-full p-5 gap-2.5 overflow-y-auto">
-        {messageHistory.map((message) => (
-          <Message key={message.id} message={message} />
-        ))}
-      </div>
-
-      {/* input */}
-      <div className="flex p-5 w-full space-x-2.5 bg-secondary text-stone-100 items-center">
-        <form
-          className="flex grow lg:mx-5 max-h-32 bg-primary opacity-70 rounded"
-          onSubmit={sendMessageHandler}
-          onKeyUp={onKeyUpHandler}
-        >
-          <textarea
-            className="grow max-h-32 shadow p-3 appearance-none text-stone-100 bg-transparent outline-none focus:outline-none focus:shadow-outline"
-            placeholder="Type a message"
-            name="message"
-            ref={messageRef}
-            type="text"
-            value={message}
-            onChange={handleChangeMessage}
-          />
-          <button
-            type="submit"
-            className="border rounded border-transparent bg-teal-300 p-3 text-primary font-medium hover:bg-teal-400"
-          >
-            Send
-          </button>
-        </form>
-      </div>
+    <div className="flex flex-col w-full lg:w-[calc(100%-500px)] h-full self-end">
+      <ChatHeader />
+      <Messages messageHistory={messageHistory} />
+      <MessageInput sendMessage={sendMessage} />
     </div>
   );
 };
